Edit path settings/profile
