+++ apps/doji-frontend/src/modules/settings/index.tsx
@@ -59,7 +59,7 @@ const SettingsPage: React.FC<SettingsPageProps> = ({ user }) => {
       </Link>
       <List component="nav" sx={{ mt: 2 }}>
         <Stack direction="column" spacing={1}>
-          <ListItem href="/settings/profile/edit" text="Edit Profile" icon={<FiEdit2 />} />
+          <ListItem href="/settings/profile" text="Edit Profile" icon={<FiEdit2 />} />
           <ListItem href="/settings/change-password" text="Change password" icon={<FiLock />} />
           <ListItem href="/settings/wallet" text="Wallet" icon={<FiDollarSign />} />
           <ListItem href="/settings/experience" text="Experience" icon={<FiLayers />} />
