+++ apps/doji-frontend/src/pages/settings/index.tsx
@@ -1,0 +1,1 @@
+export { default, getServerSideProps } from '@frontend/modules/settings'
